+++ hivemind/client/__init__.py
@@ -1,3 +1,4 @@
 from hivemind.client.expert import RemoteExpert
 from hivemind.client.moe import RemoteMixtureOfExperts
 from hivemind.client.averaging import DecentralizedAverager
+from hivemind.client.dht_ops import declare_experts, get_experts
